updated the presentation layer

+++ PresentationLayer/sadnaTeam17A/src/components/Admin/DeleteUser.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useState } from "react";
 import { ClientCredentialsContext } from "../../App";
 import axios from "axios";
-import { Button, Flex, Heading, Input, Text } from "@chakra-ui/react";
+import { Button, Flex, Heading, Input, Stack, Text } from "@chakra-ui/react";
 
 interface Props {
   setPage: React.Dispatch<React.SetStateAction<string>>;
@@ -13,31 +13,37 @@ const DeleteUser = ({ setPage, pages }: Props) => {
 
   const getUserId = async () => {
     const response = await axios.get(
-      `http://localhost:8080/api/v1/users/search-user/username=${username}`
+      `http://localhost:8080/api/v1/users/get-user-by-username/username=${username}`
     );
     if (!response.data.error) {
-      setTargetId(response.data.value[0].id);
+      setTargetId(response.data.value.id);
     } else {
+      setTargetId("");
       console.log(response.data.error);
     }
   };
 
   const handleDeleteUser = async () => {
-    const response = await axios.delete(
-      "http://localhost:8080/api/v1/users/admin/delete-user",
-      {
-        data: {
-          clientCredentials: clientCredentials,
-          targetId: targetId,
-        },
-      }
-    );
-    if (!response.data.error) {
-      setErrorMsg(false);
-      setMessage("User deleted!");
-    } else {
+    if (targetId === "") {
       setErrorMsg(true);
-      setMessage(response.data.message);
+      setMessage("User not found");
+    } else {
+      const response = await axios.delete(
+        "http://localhost:8080/api/v1/users/admin/delete-user",
+        {
+          data: {
+            clientCredentials: clientCredentials,
+            targetId: targetId,
+          },
+        }
+      );
+      if (!response.data.error) {
+        setErrorMsg(false);
+        setMessage("User deleted!");
+      } else {
+        setErrorMsg(true);
+        setMessage(response.data.message);
+      }
     }
   };
 
@@ -52,28 +58,38 @@ const DeleteUser = ({ setPage, pages }: Props) => {
 
   return (
     <>
-      <Heading padding={5} textAlign="center">
-        Delete User
-      </Heading>
-      <Input
-        bg="white"
-        placeholder="Username"
-        value={username}
-        onChange={(username) => setUsername(username.target.value)}
-      />
-      <Button colorScheme="blue" size="lg" onClick={handleDeleteUser}>
-        Delete user
-      </Button>
-      <Button
-        colorScheme="blackAlpha"
-        size="lg"
-        onClick={() => setPage(pages[0])}
-      >
-        Back
-      </Button>
-      <Flex justifyContent="center">
-        {errorMsg ? <Text color="red">{message}</Text> : <Text>{message}</Text>}
-      </Flex>
+      <Stack w="50%" maxW="400px">
+        <Heading padding={5} textAlign="center">
+          Delete User
+        </Heading>
+        <Input
+          bg="white"
+          placeholder="Username"
+          value={username}
+          onChange={(username) => {
+            setUsername(username.target.value);
+            setErrorMsg(false);
+            setMessage("");
+          }}
+        />
+        <Button colorScheme="blue" size="lg" onClick={handleDeleteUser}>
+          Delete User
+        </Button>
+        <Button
+          colorScheme="blackAlpha"
+          size="lg"
+          onClick={() => setPage(pages[0])}
+        >
+          Back
+        </Button>
+        <Flex justifyContent="center">
+          {errorMsg ? (
+            <Text color="red">{message}</Text>
+          ) : (
+            <Text>{message}</Text>
+          )}
+        </Flex>
+      </Stack>
     </>
   );
 };
